// src/app/models/Account.ts
export interface Login {
    email: string;
    password: string;
}

export interface Account {
    
    roleName:    any;
    id:          number;
    name:        string;
    email:       string;
    password:    string;
    phoneNumber: string;
    address:     string;
    image:       string;
    roleID:      number;
    role:        null;
}